Inventory에 들어갈 슬롯 추가

+++ David Engine Project/David Engine/daPlayerScript.cpp
@@ -121,7 +121,7 @@ namespace da
         if (mDashCount->MaxCount == mDashCount->CurCount)
             return;
 
-        mDashCountTime += Time::DeltaTime();
+        mDashCountTime += (float)Time::DeltaTime();
 
         if (mRegenCountTime <= mDashCountTime)
         {
